requirements.txt
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
Click==7.0
colorama==0.4.3
cycler==0.10.0
defusedxml==0.6.0
Django==3.1.1
django-allauth==0.42.0
django-cors-headers==3.5.0
django-rest-auth==0.9.5
djangorestframework==3.11.1
djangorestframework-simplejwt==4.6.0
enron-reader==0.25
entrypoints==0.3
flake8==3.7.9
idna==2.10
isort==5.5.3
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.16.0
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lxml==4.6.1
MarkupSafe==1.1.1
matplotlib==3.3.1
mccabe==0.6.1
nltk==3.5
numpy==1.19.2
oauthlib==3.1.0
Pillow==7.2.0
pycodestyle==2.6.0
pyflakes==2.1.1
PyJWT==1.7.1
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.11
sqlparse==0.3.1
toml==0.10.1
tqdm==4.49.0
urllib3==1.25.10
virtualenv==16.7.7
Werkzeug==0.16.0
wrapt==1.12.1
